+++ CHANGELOG.md
@@ -13,6 +13,9 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 ### Changed
 
 - Disabled source with name `customs.base`
+- Disabled source with name `insurance.dtp`
+- Disabled source with name `insurance.dtp.basalt`
+- Disabled source with name `insurance.registry`
 - Field `additional_info.vehicle.owner.enforcement_proceedings.has_proceedings` not fillable by `fssp.base` anymore
 - Field `tech_data.brand.name.original` not fillable by `customs.base` anymore
 - Fields `identifiers.*` not fillable by `customs.base` anymore
@@ -20,6 +23,9 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 - Fields `tech_data.*` not fillable by `customs.base` anymore
 - Fields `customs.history.*` not fillable by `customs.base` anymore
 - Fields `mileages.items[].*` not fillable by `customs.base` anymore
+- Fields `accidents.*` not fillable by `insurance.dtp` anymore
+- Fields `accidents.*` not fillable by `insurance.dtp.basalt` anymore
+- Fields `accidents.*` not fillable by `insurance.registry` anymore
 
 ## v3.161.0